# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "FundingWallResponse": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/FundingWallAsset"
            },
            "title": "Assets",
            "type": "array"
          },
          "exchanges": {
            "items": {
              "type": "string"
            },
            "title": "Exchanges",
            "type": "array"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "integer"
          }
        },
        "required": [
          "timestamp",
          "assets",
          "exchanges"
        ],
        "title": "FundingWallResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NormalizeToInterval": {
        "enum": [
          "raw",
          "1h",
          "8h",
          "1d",
          "365d"
        ],
        "title": "NormalizeToInterval",
        "type": "string"
      },
      "PaginatedCumulativeFundingDifference": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CumulativeFundingDifference"
            },
            "title": "Data",
            "type": "array"
          },
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total_count",
          "offset",
          "limit",
          "has_more"
        ],
        "title": "PaginatedCumulativeFundingDifference",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Funding Data API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v0/funding-data/diff/historical_cumulative_differences": {
      "get": {
        "operationId": "historical_cumulative_differences_api_v0_funding_data_diff_historical_cumulative_differences_get",
        "parameters": [
          {
            "in": "query",
            "name": "asset_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Names"
            }
          },
          {
            "in": "query",
            "name": "section_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section Names"
            }
          },
          {
            "in": "query",
            "name": "quote_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quote Names"
            }
          },
          {
            "in": "query",
            "name": "compare_for_section",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compare For Section"
            }
          },
          {
            "in": "query",
            "name": "min_diff",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Diff"
            }
          },
          {
            "in": "query",
            "name": "buffer_minutes",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Buffer Minutes",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "from_ts",
            "required": true,
            "schema": {
              "title": "From Ts",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "to_ts",
            "required": true,
            "schema": {
              "title": "To Ts",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCumulativeFundingDifference"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get cumulative funding differences over a time period",
        "tags": [
          "funding-data"
        ]
      }
    },
    "/api/v0/funding-data/diff/historical_differences": {
      "get": {
        "operationId": "historical_funding_differences_api_v0_funding_data_diff_historical_differences_get",
        "parameters": [
          {
            "in": "query",
            "name": "normalize_to_interval",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/NormalizeToInterval",
              "default": "raw"
            }
          },
          {
            "in": "query",
            "name": "asset_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Names"
            }
          },
          {
            "in": "query",
            "name": "section_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section Names"
            }
          },
          {
            "in": "query",
            "name": "quote_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quote Names"
            }
          },
          {
            "in": "query",
            "name": "compare_for_section",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compare For Section"
            }
          },
          {
            "in": "query",
            "name": "min_diff",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Diff"
            }
          },
          {
            "in": "query",
            "name": "buffer_minutes",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Buffer Minutes",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "from_ts",
            "required": true,
            "schema": {
              "title": "From Ts",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "to_ts",
            "required": true,
            "schema": {
              "title": "To Ts",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCumulativeFundingDifference"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get historical funding differences over a time period with optional normalization",
        "tags": [
          "funding-data"
        ]
      }
    },
...
    "/api/v0/funding-data/funding-wall": {
      "get": {
        "operationId": "funding_wall_api_v0_funding_data_funding_wall_get",
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "required": true,
            "schema": {
              "enum": [
                "live",
                "historical"
              ],
              "title": "Period",
              "type": "string"
            }
          },
          {
            "description": "List of asset names to filter by",
            "in": "query",
            "name": "assets",
            "required": true,
            "schema": {
              "description": "List of asset names to filter by",
              "items": {
                "type": "string"
              },
              "title": "Assets",
              "type": "array"
            }
          },
          {
            "description": "List of exchange names to filter by",
            "in": "query",
            "name": "exchanges",
            "required": true,
            "schema": {
              "description": "List of exchange names to filter by",
              "items": {
                "type": "string"
              },
              "title": "Exchanges",
              "type": "array"
            }
          },
          {
            "description": "Normalization interval",
            "in": "query",
            "name": "normalize",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/NormalizeToInterval",
              "description": "Normalization interval"
            }
          },
          {
            "in": "query",
            "name": "from_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Ts"
            }
          },
          {
            "in": "query",
            "name": "to_ts",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Ts"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundingWallResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get funding wall matrix data",
        "tags": [
          "funding-data"
        ]
      }
    },
    "/api/v0/funding-data/historical_avg": {
      "get": {
        "operationId": "historical_avg_api_v0_funding_data_historical_avg_get",
        "parameters": [
          {
            "in": "query",
            "name": "normalize_to_interval",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/NormalizeToInterval",
              "default": "365d"
            }
          },
          {
            "in": "query",
            "name": "asset_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Names"
            }
          },
          {
            "in": "query",
            "name": "section_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section Names"
            }
          },
          {
            "in": "query",
            "name": "quote_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quote Names"
            }
          },
          {
            "description": "Window sizes in days, e.g. 7 30 90",
            "in": "query",
            "name": "windows",
            "required": true,
            "schema": {
              "description": "Window sizes in days, e.g. 7 30 90",
              "items": {
                "type": "integer"
              },
              "title": "Windows",
              "type": "array"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoricalAvgEntry"
                  },
                  "title": "Response Historical Avg Api V0 Funding Data Historical Avg Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Average historical funding per contract over one or more day windows",
        "tags": [
          "funding-data"
        ]
      }
    },
    "/api/v0/funding-data/historical_latest": {
      "get": {
        "operationId": "historical_latest_api_v0_funding_data_historical_latest_get",
        "parameters": [
          {
            "in": "query",
            "name": "normalize_to_interval",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/NormalizeToInterval",
              "default": "365d"
            }
          },
          {
            "in": "query",
            "name": "asset_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Asset Names"
            }
          },
          {
            "in": "query",
            "name": "section_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section Names"
            }
          },
          {
            "in": "query",
            "name": "quote_names",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quote Names"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LatestFundingPoint"
                  },
                  "title": "Response Historical Latest Api V0 Funding Data Historical Latest Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Latest settled historical funding rate per contract for a filter slice",
        "tags": [
          "funding-data"
        ]
      }
    },
...
    "/api/v0/funding-data/period_sums/{contract_id}": {
      "get": {
        "operationId": "period_sums_api_v0_funding_data_period_sums__contract_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "contract_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Contract Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundingPeriodSums"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get funding sums for different time periods for a specific contract",
        "tags": [
          "funding-data"
        ]
      }
    },
...
  }
}